view-edit: Improved <label> elements for complex settings markup.
A number of settings on view-edit contain multiple input/select elements.
Some of these elements didn't have labels. This changeset rearranges some
markup so that labels are applied in ways that make semantic sense, mostly
using the text/markup that is already there.

As part of this commit, a checkbox with `visibility:hidden;` is also removed.
It was added in e120e5e7dffd7838793b7785294ac66aaa42c61e to ensure proper
spacing of the second line of the settings text. But this causes problems
for accessibility tools that can't be easily avoided using ARIA roles, so
I've swapped the technique for pure CSS.