Unit tests: Update PHPUnit bootstrap.
The "old" method of hoisting the plugin loader file into the
`wp_tests_options` global doesn't work in all versions of WordPress.
Instead, we manually load the plugin file in a `tests_add_filter()`
callback. The `wp_tests_options` modification is kept in place to
ensure that TablePress's "is plugin active" test continues to pass.

This change requires that some PHP constants be defined conditionally,
to avoid PHP notices on Travis CI.