Remove the migration and import code for the outdated WP-Table Reloaded plugin.
TablePress has been around for over 7 years now, and since then, WP-Table Reloaded has contained a nagging warning that people should switch. The 30000 currently still active installations of WP-Table Reloaded are probably on old and unmaintained sites anyways, so that it's unlikely that they will migrate soon. If they want to, they can simply add an intermediate step and install an older version of TablePress, like 1.9.2, potentially on an older version of WordPress, even on a dev site, and migrate to the TablePress JSON data format there.

For TablePress, this removes a big chunk of code that now won't need to be maintained anymore, which makes the plugin more lightweight.